split a new file named sarama.go
