+++ pkg/sink/kafka/options_test.go
@@ -1,4 +1,4 @@
-// Copyright 2021 PingCAP, Inc.
+// Copyright 2023 PingCAP, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,110 +14,18 @@
 package kafka
 
 import (
-	"context"
 	"fmt"
 	"net/url"
 	"testing"
 	"time"
 
-	"github.com/Shopify/sarama"
 	"github.com/pingcap/errors"
-	"github.com/pingcap/tiflow/cdc/contextutil"
 	"github.com/pingcap/tiflow/cdc/model"
 	cerror "github.com/pingcap/tiflow/pkg/errors"
-	"github.com/pingcap/tiflow/pkg/security"
 	"github.com/stretchr/testify/require"
 )
 
-func TestNewSaramaConfig(t *testing.T) {
-	ctx := context.Background()
-	options := NewOptions()
-	options.Version = "invalid"
-	_, err := NewSaramaConfig(ctx, options)
-	require.Regexp(t, "invalid version.*", errors.Cause(err))
-	ctx = contextutil.SetOwnerInCtx(ctx)
-	options.Version = "2.6.0"
-	options.ClientID = "^invalid$"
-	_, err = NewSaramaConfig(ctx, options)
-	require.True(t, cerror.ErrKafkaInvalidClientID.Equal(err))
-
-	options.ClientID = "test-kafka-client"
-	compressionCases := []struct {
-		algorithm string
-		expected  sarama.CompressionCodec
-	}{
-		{"none", sarama.CompressionNone},
-		{"gzip", sarama.CompressionGZIP},
-		{"snappy", sarama.CompressionSnappy},
-		{"lz4", sarama.CompressionLZ4},
-		{"zstd", sarama.CompressionZSTD},
-		{"others", sarama.CompressionNone},
-	}
-	for _, cc := range compressionCases {
-		options.Compression = cc.algorithm
-		cfg, err := NewSaramaConfig(ctx, options)
-		require.Nil(t, err)
-		require.Equal(t, cc.expected, cfg.Producer.Compression)
-	}
-
-	options.EnableTLS = true
-	options.Credential = &security.Credential{
-		CAPath:   "/invalid/ca/path",
-		CertPath: "/invalid/cert/path",
-		KeyPath:  "/invalid/key/path",
-	}
-	_, err = NewSaramaConfig(ctx, options)
-	require.Regexp(t, ".*no such file or directory", errors.Cause(err))
-
-	saslOptions := NewOptions()
-	saslOptions.Version = "2.6.0"
-	saslOptions.ClientID = "test-sasl-scram"
-	saslOptions.SASL = &security.SASL{
-		SASLUser:      "user",
-		SASLPassword:  "password",
-		SASLMechanism: sarama.SASLTypeSCRAMSHA256,
-	}
-
-	cfg, err := NewSaramaConfig(ctx, saslOptions)
-	require.Nil(t, err)
-	require.NotNil(t, cfg)
-	require.Equal(t, "user", cfg.Net.SASL.User)
-	require.Equal(t, "password", cfg.Net.SASL.Password)
-	require.Equal(t, sarama.SASLMechanism("SCRAM-SHA-256"), cfg.Net.SASL.Mechanism)
-}
-
-func TestConfigTimeouts(t *testing.T) {
-	options := NewOptions()
-	require.Equal(t, 10*time.Second, options.DialTimeout)
-	require.Equal(t, 10*time.Second, options.ReadTimeout)
-	require.Equal(t, 10*time.Second, options.WriteTimeout)
-
-	saramaConfig, err := NewSaramaConfig(context.Background(), options)
-	require.Nil(t, err)
-	require.Equal(t, options.DialTimeout, saramaConfig.Net.DialTimeout)
-	require.Equal(t, options.WriteTimeout, saramaConfig.Net.WriteTimeout)
-	require.Equal(t, options.ReadTimeout, saramaConfig.Net.ReadTimeout)
-
-	uri := "kafka://127.0.0.1:9092/kafka-test?dial-timeout=5s&read-timeout=1000ms" +
-		"&write-timeout=2m"
-	sinkURI, err := url.Parse(uri)
-	require.Nil(t, err)
-
-	err = options.Apply(sinkURI)
-	require.Nil(t, err)
-
-	require.Equal(t, 5*time.Second, options.DialTimeout)
-	require.Equal(t, 1000*time.Millisecond, options.ReadTimeout)
-	require.Equal(t, 2*time.Minute, options.WriteTimeout)
-
-	saramaConfig, err = NewSaramaConfig(context.Background(), options)
-	require.Nil(t, err)
-	require.Equal(t, 5*time.Second, saramaConfig.Net.DialTimeout)
-	require.Equal(t, 1000*time.Millisecond, saramaConfig.Net.ReadTimeout)
-	require.Equal(t, 2*time.Minute, saramaConfig.Net.WriteTimeout)
-}
-
-func TestCompleteConfigByOpts(t *testing.T) {
+func TestCompleteOptions(t *testing.T) {
 	options := NewOptions()
 
 	// Normal config.
@@ -127,10 +35,10 @@ func TestCompleteConfigByOpts(t *testing.T) {
 	maxMessageSize := "4096" // 4kb
 	uri := fmt.Sprintf(uriTemplate, maxMessageSize)
 	sinkURI, err := url.Parse(uri)
-	require.Nil(t, err)
+	require.NoError(t, err)
 
 	err = options.Apply(sinkURI)
-	require.Nil(t, err)
+	require.NoError(t, err)
 	require.Equal(t, int32(1), options.PartitionNum)
 	require.Equal(t, int16(3), options.ReplicationFactor)
 	require.Equal(t, "2.6.0", options.Version)
@@ -139,16 +47,16 @@ func TestCompleteConfigByOpts(t *testing.T) {
 	// multiple kafka broker endpoints
 	uri = "kafka://127.0.0.1:9092,127.0.0.1:9091,127.0.0.1:9090/kafka-test?"
 	sinkURI, err = url.Parse(uri)
-	require.Nil(t, err)
+	require.NoError(t, err)
 	options = NewOptions()
 	err = options.Apply(sinkURI)
-	require.Nil(t, err)
+	require.NoError(t, err)
 	require.Len(t, options.BrokerEndpoints, 3)
 
 	// Illegal replication-factor.
 	uri = "kafka://127.0.0.1:9092/abc?kafka-version=2.6.0&replication-factor=a"
 	sinkURI, err = url.Parse(uri)
-	require.Nil(t, err)
+	require.NoError(t, err)
 	options = NewOptions()
 	err = options.Apply(sinkURI)
 	require.Regexp(t, ".*invalid syntax.*", errors.Cause(err))
@@ -156,7 +64,7 @@ func TestCompleteConfigByOpts(t *testing.T) {
 	// Illegal max-message-bytes.
 	uri = "kafka://127.0.0.1:9092/abc?kafka-version=2.6.0&max-message-bytes=a"
 	sinkURI, err = url.Parse(uri)
-	require.Nil(t, err)
+	require.NoError(t, err)
 	options = NewOptions()
 	err = options.Apply(sinkURI)
 	require.Regexp(t, ".*invalid syntax.*", errors.Cause(err))
@@ -164,7 +72,7 @@ func TestCompleteConfigByOpts(t *testing.T) {
 	// Illegal partition-num.
 	uri = "kafka://127.0.0.1:9092/abc?kafka-version=2.6.0&partition-num=a"
 	sinkURI, err = url.Parse(uri)
-	require.Nil(t, err)
+	require.NoError(t, err)
 	options = NewOptions()
 	err = options.Apply(sinkURI)
 	require.Regexp(t, ".*invalid syntax.*", errors.Cause(err))
@@ -172,7 +80,7 @@ func TestCompleteConfigByOpts(t *testing.T) {
 	// Out of range partition-num.
 	uri = "kafka://127.0.0.1:9092/abc?kafka-version=2.6.0&partition-num=0"
 	sinkURI, err = url.Parse(uri)
-	require.Nil(t, err)
+	require.NoError(t, err)
 	options = NewOptions()
 	err = options.Apply(sinkURI)
 	require.Regexp(t, ".*invalid partition num.*", errors.Cause(err))
@@ -181,190 +89,17 @@ func TestCompleteConfigByOpts(t *testing.T) {
 func TestSetPartitionNum(t *testing.T) {
 	options := NewOptions()
 	err := options.SetPartitionNum(2)
-	require.Nil(t, err)
+	require.NoError(t, err)
 	require.Equal(t, int32(2), options.PartitionNum)
 
 	options.PartitionNum = 1
 	err = options.SetPartitionNum(2)
-	require.Nil(t, err)
+	require.NoError(t, err)
 	require.Equal(t, int32(1), options.PartitionNum)
 
 	options.PartitionNum = 3
 	err = options.SetPartitionNum(2)
 	require.True(t, cerror.ErrKafkaInvalidPartitionNum.Equal(err))
-}
-
-func TestApplySASL(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name      string
-		URI       string
-		exceptErr string
-	}{
-		{
-			name:      "no params",
-			URI:       "kafka://127.0.0.1:9092/abc",
-			exceptErr: "",
-		},
-		{
-			name: "valid PLAIN SASL",
-			URI: "kafka://127.0.0.1:9092/abc?kafka-version=2.6.0&partition-num=0" +
-				"&sasl-user=user&sasl-password=password&sasl-mechanism=plain",
-			exceptErr: "",
-		},
-		{
-			name: "valid SCRAM SASL",
-			URI: "kafka://127.0.0.1:9092/abc?kafka-version=2.6.0&partition-num=0" +
-				"&sasl-user=user&sasl-password=password&sasl-mechanism=SCRAM-SHA-512",
-			exceptErr: "",
-		},
-		{
-			name: "valid GSSAPI user auth SASL",
-			URI: "kafka://127.0.0.1:9092/abc?kafka-version=2.6.0&partition-num=0" +
-				"&sasl-mechanism=GSSAPI&sasl-gssapi-auth-type=USER" +
-				"&sasl-gssapi-kerberos-config-path=/root/config" +
-				"&sasl-gssapi-service-name=a&sasl-gssapi-user=user" +
-				"&sasl-gssapi-password=pwd" +
-				"&sasl-gssapi-realm=realm&sasl-gssapi-disable-pafxfast=false",
-			exceptErr: "",
-		},
-		{
-			name: "valid GSSAPI keytab auth SASL",
-			URI: "kafka://127.0.0.1:9092/abc?kafka-version=2.6.0&partition-num=0" +
-				"&sasl-mechanism=GSSAPI&sasl-gssapi-auth-type=keytab" +
-				"&sasl-gssapi-kerberos-config-path=/root/config" +
-				"&sasl-gssapi-service-name=a&sasl-gssapi-user=user" +
-				"&sasl-gssapi-keytab-path=/root/keytab" +
-				"&sasl-gssapi-realm=realm&sasl-gssapi-disable-pafxfast=false",
-			exceptErr: "",
-		},
-		{
-			name: "invalid mechanism",
-			URI: "kafka://127.0.0.1:9092/abc?kafka-version=2.6.0&partition-num=0" +
-				"&sasl-mechanism=a",
-			exceptErr: "unknown a SASL mechanism",
-		},
-		{
-			name: "invalid GSSAPI auth type",
-			URI: "kafka://127.0.0.1:9092/abc?kafka-version=2.6.0&partition-num=0" +
-				"&sasl-mechanism=gssapi&sasl-gssapi-auth-type=keyta1b",
-			exceptErr: "unknown keyta1b auth type",
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			t.Parallel()
-			options := NewOptions()
-			sinkURI, err := url.Parse(test.URI)
-			require.Nil(t, err)
-			if test.exceptErr == "" {
-				require.Nil(t, options.applySASL(sinkURI.Query()))
-			} else {
-				require.Regexp(t, test.exceptErr, options.applySASL(sinkURI.Query()).Error())
-			}
-		})
-	}
-}
-
-func TestApplyTLS(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name       string
-		URI        string
-		tlsEnabled bool
-		exceptErr  string
-	}{
-		{
-			name: "tls config with 'enable-tls' set to true",
-			URI: "kafka://127.0.0.1:9092/abc?kafka-version=2.6.0&partition-num=0" +
-				"&sasl-user=user&sasl-password=password&sasl-mechanism=plain&enable-tls=true",
-			tlsEnabled: true,
-			exceptErr:  "",
-		},
-		{
-			name: "tls config with no 'enable-tls', and credential files are supplied",
-			URI: "kafka://127.0.0.1:9092/abc?kafka-version=2.6.0&partition-num=0" +
-				"&sasl-user=user&sasl-password=password&sasl-mechanism=plain" +
-				"&ca=/root/ca.file&cert=/root/cert.file&key=/root/key.file",
-			tlsEnabled: true,
-			exceptErr:  "",
-		},
-		{
-			name: "tls config with no 'enable-tls', and credential files are not supplied",
-			URI: "kafka://127.0.0.1:9092/abc?kafka-version=2.6.0&partition-num=0" +
-				"&sasl-user=user&sasl-password=password&sasl-mechanism=plain",
-			tlsEnabled: false,
-			exceptErr:  "",
-		},
-		{
-			name: "tls config with 'enable-tls' set to false, and credential files are supplied",
-			URI: "kafka://127.0.0.1:9092/abc?kafka-version=2.6.0&partition-num=0" +
-				"&sasl-user=user&sasl-password=password&sasl-mechanism=plain&enable-tls=false" +
-				"&ca=/root/ca&cert=/root/cert&key=/root/key",
-			tlsEnabled: false,
-			exceptErr:  "credential files are supplied, but 'enable-tls' is set to false",
-		},
-		{
-			name: "tls config with 'enable-tls' set to true, and some of " +
-				"the credential files are not supplied ",
-			URI: "kafka://127.0.0.1:9092/abc?kafka-version=2.6.0&partition-num=0" +
-				"&sasl-user=user&sasl-password=password&sasl-mechanism=plain&enable-tls=true" +
-				"&ca=/root/ca&cert=/root/cert&",
-			tlsEnabled: false,
-			exceptErr:  "ca, cert and key files should all be supplied",
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			t.Parallel()
-			options := NewOptions()
-			sinkURI, err := url.Parse(test.URI)
-			require.Nil(t, err)
-			if test.exceptErr == "" {
-				require.Nil(t, options.applyTLS(sinkURI.Query()))
-			} else {
-				require.Regexp(t, test.exceptErr, options.applyTLS(sinkURI.Query()).Error())
-			}
-			require.Equal(t, test.tlsEnabled, options.EnableTLS)
-		})
-	}
-}
-
-func TestCompleteSaramaSASLConfig(t *testing.T) {
-	t.Parallel()
-
-	// Test that SASL is turned on correctly.
-	options := NewOptions()
-	options.SASL = &security.SASL{
-		SASLUser:      "user",
-		SASLPassword:  "password",
-		SASLMechanism: "",
-		GSSAPI:        security.GSSAPI{},
-	}
-	saramaConfig := sarama.NewConfig()
-	completeSaramaSASLConfig(saramaConfig, options)
-	require.False(t, saramaConfig.Net.SASL.Enable)
-	options.SASL.SASLMechanism = "plain"
-	completeSaramaSASLConfig(saramaConfig, options)
-	require.True(t, saramaConfig.Net.SASL.Enable)
-	// Test that the SCRAMClientGeneratorFunc is set up correctly.
-	options = NewOptions()
-	options.SASL = &security.SASL{
-		SASLUser:      "user",
-		SASLPassword:  "password",
-		SASLMechanism: "plain",
-		GSSAPI:        security.GSSAPI{},
-	}
-	saramaConfig = sarama.NewConfig()
-	completeSaramaSASLConfig(saramaConfig, options)
-	require.Nil(t, saramaConfig.Net.SASL.SCRAMClientGeneratorFunc)
-	options.SASL.SASLMechanism = "SCRAM-SHA-512"
-	completeSaramaSASLConfig(saramaConfig, options)
-	require.NotNil(t, saramaConfig.Net.SASL.SCRAMClientGeneratorFunc)
 }
 
 func TestClientID(t *testing.T) {
@@ -407,8 +142,27 @@ func TestClientID(t *testing.T) {
 		if tc.hasError {
 			require.Error(t, err)
 		} else {
-			require.Nil(t, err)
+			require.NoError(t, err)
 			require.Equal(t, tc.expected, id)
 		}
 	}
+}
+
+func TestTimeout(t *testing.T) {
+	options := NewOptions()
+	require.Equal(t, 10*time.Second, options.DialTimeout)
+	require.Equal(t, 10*time.Second, options.ReadTimeout)
+	require.Equal(t, 10*time.Second, options.WriteTimeout)
+
+	uri := "kafka://127.0.0.1:9092/kafka-test?dial-timeout=5s&read-timeout=1000ms" +
+		"&write-timeout=2m"
+	sinkURI, err := url.Parse(uri)
+	require.NoError(t, err)
+
+	err = options.Apply(sinkURI)
+	require.NoError(t, err)
+
+	require.Equal(t, 5*time.Second, options.DialTimeout)
+	require.Equal(t, 1000*time.Millisecond, options.ReadTimeout)
+	require.Equal(t, 2*time.Minute, options.WriteTimeout)
 }
